src/modules/users/services/GetUserProfileService.ts
import { injectable, inject } from "tsyringe";

import AppError from "@shared/errors/AppError";
import User from "../infra/typeorm/entities/User";
import IUsersRepository from "../repositories/IUsersRepository";

@injectable()
class GetUserProfileService {
  constructor(
    @inject("UsersRepository")
    private readonly usersRepository: IUsersRepository
  ) {}

  public async execute(userId: string): Promise<User> {
    const user = await this.usersRepository.findById(userId);

    if (!user) {
      throw new AppError("An user with this id was not found.", 404);
    }

    return user;
  }
}

export default GetUserProfileService;
